+++ iivs/dhm/data/intensity/convert.py
@@ -6,6 +6,7 @@ from functools import partial
 from typing import TYPE_CHECKING
 
 from kaparoo.filesystem import StagedDirectory
+from kaparoo.utils import ensure_one_of
 
 from iivs.dhm.data.common import numbered_name
 from iivs.dhm.data.intensity.bin import save_intensity_bin
@@ -45,9 +46,7 @@ def convert_intensity_folder(
         ValueError: If `ext` is not "bin", "txt", or "npy".
         FileExistsError: If `root` exists and `overwrite` is False.
     """
-    if ext not in ("bin", "txt", "npy"):
-        msg = f"ext must be 'bin', 'txt', or 'npy' (got {ext!r})"
-        raise ValueError(msg)
+    ensure_one_of(ext, ("bin", "txt", "npy"), name="ext")
 
     if ext == "npy":
         save = partial(save_intensity_npy, overwrite=overwrite)
@@ -84,9 +83,7 @@ def convert_intensity_list(
         ValueError: If `ext` is not "bin", "txt", or "npy".
         FileExistsError: If a target sibling exists and `overwrite` is False.
     """
-    if ext not in ("bin", "txt", "npy"):
-        msg = f"ext must be 'bin', 'txt', or 'npy' (got {ext!r})"
-        raise ValueError(msg)
+    ensure_one_of(ext, ("bin", "txt", "npy"), name="ext")
 
     if ext == "npy":
         save = partial(save_intensity_npy, overwrite=overwrite)